Warning cleanup and clear newly allocated klog buffers

+++ kernel/arch/x86/include/arch/pit.h
@@ -6,8 +6,8 @@
 void pit_init();
 
 // Deprecated interface
-void pit_init_counter(uint16_t id, uint16_t frequency, uint8_t mode);
+void pit_init_counter(uint16_t id, uint32_t frequency, uint8_t mode);
 void pit_reset_counter(uint16_t id);
-void pit_set_counter(uint16_t id, uint16_t frequency);
+void pit_set_counter(uint16_t id, uint32_t frequency);
 
 #endif /* __PIT_H__ */
